+++ pid.h
@@ -30,19 +30,32 @@
  */
 typedef struct PID_DATA{
   //! Last process value, used to find derivative of process value.
-  int16_t lastProcessValue;
+//  int16_t lastProcessValue;
+  double lastProcessValue;
+
   //! Summation of errors, used for integrate calculations
-  int32_t sumError;
+//  int32_t sumError;
+  double sumError;
+
   //! The Proportional tuning constant, multiplied with SCALING_FACTOR
-  int16_t P_Factor;
+//  int16_t P_Factor;
+  double P_Factor;
+
   //! The Integral tuning constant, multiplied with SCALING_FACTOR
-  int16_t I_Factor;
+//  int16_t I_Factor;
+  double I_Factor;
+
   //! The Derivative tuning constant, multiplied with SCALING_FACTOR
-  int16_t D_Factor;
+//  int16_t D_Factor;
+  double D_Factor;
+
   //! Maximum allowed error, avoid overflow
-  int16_t maxError;
+//  int16_t maxError;
+  double maxError;
+
   //! Maximum allowed sumerror, avoid overflow
-  int32_t maxSumError;
+//  int32_t maxSumError;
+  double maxSumError;
 } pidData_t;
 
 /*! \brief Maximum values
@@ -58,9 +71,12 @@ typedef struct PID_DATA{
 //#define FALSE           0
 //#define TRUE            1
 
-void pid_Init(int16_t p_factor, int16_t i_factor, int16_t d_factor, struct PID_DATA *pid);
+//void pid_Init(int16_t p_factor, int16_t i_factor, int16_t d_factor, struct PID_DATA *pid);
+void pid_Init(double p_factor, double i_factor, double d_factor, struct PID_DATA *pid);
 
-int16_t pid_Controller(int16_t setPoint, int16_t processValue, struct PID_DATA *pid_st);
+
+//int16_t pid_Controller(int16_t setPoint, int16_t processValue, struct PID_DATA *pid_st);
+int16_t pid_Controller(double setPoint, double processValue, struct PID_DATA *pid_st);
 
 void pid_Reset_Integrator(pidData_t *pid_st);
